fix Int to BigInt to store nanoseconds

+++ LoRaTimeSyncServerApp/models.py
@@ -4,15 +4,15 @@ from django.db import models
 class TimeSyncInit(models.Model):
     dev_eui = models.CharField(max_length=16)
     period = models.IntegerField()
-    first_uplink_expected = models.IntegerField()
+    first_uplink_expected = models.BigIntegerField()
     created_at = models.DateTimeField(auto_now_add=True)
 
 
 class TimeCollection(models.Model):
     dev_eui = models.CharField(max_length=16)
     device_time = models.IntegerField()
-    time_expected = models.IntegerField()
-    time_received = models.IntegerField()
+    time_expected = models.BigIntegerField()
+    time_received = models.BigIntegerField()
 
 
 class TimeSyncModels(models.Model):
